Moved CSVParser to utils/, final housekeeping, clang-format
