+++ quisp/utils/CSVParser.cc
@@ -3,23 +3,24 @@
  *
  *  Created on: Nov 28, 2023
  *      Author: paolo
+ *  Quick parser for CSVs, written for csv-based Free-space channels.
  */
 
 #include "CSVParser.h"
 
 CSVParser::CSVParser(const string filename) {
-    file.open(filename);
-    if (!file.is_open()) throw cRuntimeError("Couldn't find CSV file!");
-    string line;
-    double key = 0;
-    double val = 0;
-    char sep; //throwaway
-    while (file >> line) {
-        std::istringstream ss(line);
-        ss >> key >> sep >> val;
-        property.insert(std::pair<double,double>(key,val));
-    }
-    return;
+  file.open(filename);
+  if (!file.is_open()) throw cRuntimeError("Couldn't find CSV file!");
+  string line;
+  double key = 0;
+  double val = 0;
+  char sep;  // throwaway
+  while (file >> line) {
+    std::istringstream ss(line);
+    ss >> key >> sep >> val;
+    property.insert(std::pair<double, double>(key, val));
+  }
+  return;
 }
 
 double CSVParser::getPropertyAtTime(const double time) {
@@ -47,21 +48,12 @@ double CSVParser::getPropertyAtTime(const double time) {
   return last_polled_value;
 }
 
+double CSVParser::getLowestDatapoint() { return property.begin()->first; }
 
-double CSVParser::getLowestDatapoint() {
-    return property.begin()->first;
-}
-
-double CSVParser::getHighestDatapoint() {
-    return property.rbegin()->first;
-}
+double CSVParser::getHighestDatapoint() { return property.rbegin()->first; }
 
 double CSVParser::getLowestDatavalue() { return property.begin()->second; }
 
 double CSVParser::getHighestDatavalue() { return property.rbegin()->second; }
 
-
-
-CSVParser::~CSVParser() {
-}
-
+CSVParser::~CSVParser() {}
